Remove redundant log redirect
Two exec's were redirecting their output into a log file. Turns out
this is redundant as Puppet can capture the output for us and return
the result when something goes wrong.

Also this fixes a regression introduced in the previous release due
to which the database update exec lost permissions to write into the
log file.